backend/requirements.txt
# API Framework
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.9
starlette>=0.37.2,<0.38.0

# Database
redis==5.0.1
aioredis==2.0.1
supabase==1.2.0
gotrue==1.1.0
postgrest>=0.10.8,<0.12.0

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
cryptography==42.0.8
pyjwt==2.10.1

# Data Processing
pydantic==2.6.4
pandas==2.2.0
numpy==1.26.0
python-dateutil==2.8.2
tzdata==2024.2

# Testing & Development
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
httpx>=0.24.0,<0.26.0

# Utilities
requests==2.31.0
aiohttp==3.9.3
websockets==12.0
python-socketio==5.11.1
typer==0.9.0
rich==13.7.1
structlog==24.1.0
prometheus-client==0.20.0

# Game-specific
pygame==2.5.2
numba==0.59.1
scipy==1.12.0
